Change argument type from nom::uint32 to float

+++ include/nomlib/system/GameStates.hpp
@@ -47,7 +47,7 @@ class GameStates
     static void onEvent ( void* );
 
     /// State logic
-    static void Update ( uint32 delta_time );
+    static void Update ( float delta_time );
 
     /// State rendering
     static void Draw ( void* video_buffer );
